recklessly sprinkle typecasts all over the novte stuff, this should remove those annoying warnings about 64 to 32 bit truncation. please tests this
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@45669 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/gtk3/novte/iso2022.cc
@@ -170,7 +170,7 @@ gsize _vte_iso2022_process(struct _vte_iso2022_state *state,
 
     /* skip blanks -- TODOegmont: why here? */
     j = gunichars->len;
-    g_array_set_size(gunichars, gunichars->len + outbuf-buf);
+    g_array_set_size(gunichars, (guint)(gunichars->len + (outbuf - buf)));
     for (i = 0; buf + i < outbuf; i++) {
         c = buf[i];
         if (G_UNLIKELY (c == '\0')) {
